refactor(content): collapse iframe module to single file, add dom-tree entry points, remove iframe debug logs, and drop content unit tests
