+++ content/src/html.rs
@@ -1,10 +1,11 @@
 mod environment_settings_object;
 mod global_scope;
 mod html_anchor_element;
+mod html_dom_tree;
 mod html_element;
+mod html_iframe_element;
 mod html_parser;
 mod hyperlink_element_utils;
-mod iframe;
 mod window;
 mod window_or_worker_global_scope;
 
@@ -12,8 +13,12 @@ pub use environment_settings_object::EnvironmentSettingsObject;
 pub(crate) use global_scope::TimerHandler;
 pub use global_scope::{GlobalScope, GlobalScopeKind};
 pub use html_anchor_element::HTMLAnchorElement;
+pub(crate) use html_dom_tree::{
+    run_dom_post_connection_steps_for_document, run_dom_removing_steps_for_document,
+};
 pub use html_element::HTMLElement;
-pub use iframe::HTMLIFrameElement;
+pub use html_iframe_element::HTMLIFrameElement;
+pub(crate) use html_iframe_element::run_iframe_load_event_steps_for_traversable;
 pub(crate) use html_parser::PendingParserScript;
 pub use html_parser::{JsHtmlParserProvider, execute_parser_scripts, parse_html_into_document};
 pub(crate) use hyperlink_element_utils::HyperlinkElementUtils;
