+++ content/src/html/html_parser.rs
@@ -336,76 +336,3 @@ fn parser_script_for_node(document: &BaseDocument, node_id: usize) -> Option<Pen
     Some(PendingParserScript::Inline { source })
 }
 
-#[cfg(test)]
-mod tests {
-    use super::{PendingParserScript, execute_parser_scripts, parse_html_into_document};
-    use crate::html::EnvironmentSettingsObject;
-    use blitz_dom::{BaseDocument, DocumentConfig};
-    use std::{cell::RefCell, rc::Rc};
-    use url::Url;
-
-    #[test]
-    fn parse_html_into_document_skips_non_classic_script_types() {
-        let document = Rc::new(RefCell::new(BaseDocument::new(DocumentConfig::default())));
-        let scripts = {
-            let mut document_guard = document.borrow_mut();
-            parse_html_into_document(
-                &mut document_guard,
-                r#"
-                <script type="application/json">{"a": 1}</script>
-                <script type="speculationrules">{"prefetch": []}</script>
-                <script type="module">globalThis.moduleLoaded = true;</script>
-                <script>globalThis.inlineClassic = true;</script>
-                <script type="text/javascript">globalThis.typedClassic = true;</script>
-                "#,
-            )
-        };
-
-        assert_eq!(
-            scripts.len(),
-            2,
-            "expected only classic scripts to be queued"
-        );
-        assert!(matches!(
-            &scripts[0],
-            PendingParserScript::Inline { source } if source.contains("inlineClassic")
-        ));
-        assert!(matches!(
-            &scripts[1],
-            PendingParserScript::Inline { source } if source.contains("typedClassic")
-        ));
-    }
-
-    #[test]
-    fn execute_parser_scripts_keeps_loading_after_script_error() {
-        let document = Rc::new(RefCell::new(BaseDocument::new(DocumentConfig::default())));
-        let mut settings = EnvironmentSettingsObject::new(
-            Rc::clone(&document),
-            Url::parse("https://example.com/").expect("test URL should parse"),
-        )
-        .expect("environment settings should initialize");
-
-        let result = execute_parser_scripts(
-            &mut settings,
-            vec![
-                PendingParserScript::Inline {
-                    source: String::from("not valid javascript: {"),
-                },
-                PendingParserScript::Inline {
-                    source: String::from("globalThis.afterError = 1;"),
-                },
-            ],
-        );
-
-        assert!(
-            result.is_ok(),
-            "parser-script errors should not abort document loading"
-        );
-        let after_error = settings
-            .context
-            .global_object()
-            .get(boa_engine::js_string!("afterError"), &mut settings.context)
-            .expect("follow-up script should be readable");
-        assert_eq!(after_error.as_number(), Some(1.0));
-    }
-}
