--- content/src/html/iframe/element.rs
@@ -1,97 +1,0 @@
-use std::{cell::RefCell, rc::Rc};
-
-use blitz_dom::BaseDocument;
-use boa_engine::JsData;
-use boa_gc::{Finalize, Trace};
-
-use crate::html::HTMLElement;
-
-/// <https://html.spec.whatwg.org/#htmliframeelement>
-#[derive(Trace, Finalize, JsData)]
-pub struct HTMLIFrameElement {
-    /// <https://html.spec.whatwg.org/#htmlelement>
-    pub html_element: HTMLElement,
-}
-
-impl HTMLIFrameElement {
-    pub fn new(document: Rc<RefCell<BaseDocument>>, node_id: usize) -> Self {
-        Self {
-            html_element: HTMLElement::new(document, node_id),
-        }
-    }
-
-    /// <https://html.spec.whatwg.org/#dom-iframe-src>
-    pub(crate) fn src(&self) -> String {
-        // Step 1: "Return the value of the src content attribute."
-        self.html_element
-            .element
-            .get_attribute("src")
-            .unwrap_or_default()
-    }
-
-    /// <https://html.spec.whatwg.org/#dom-iframe-src>
-    pub(crate) fn set_src(&self, src: &str) {
-        // Step 1: "Set this's src content attribute to the given value."
-        self.html_element.element.set_attribute("src", src);
-    }
-
-    /// <https://html.spec.whatwg.org/#dom-iframe-srcdoc>
-    pub(crate) fn srcdoc(&self) -> String {
-        // Step 1: "Return the value of the srcdoc content attribute."
-        self.html_element
-            .element
-            .get_attribute("srcdoc")
-            .unwrap_or_default()
-    }
-
-    /// <https://html.spec.whatwg.org/#dom-iframe-srcdoc>
-    pub(crate) fn set_srcdoc(&self, srcdoc: &str) {
-        // Step 1: "Set this's srcdoc content attribute to the given value."
-        self.html_element.element.set_attribute("srcdoc", srcdoc);
-    }
-
-    /// <https://html.spec.whatwg.org/#dom-iframe-name>
-    pub(crate) fn name(&self) -> String {
-        // Step 1: "Return the value of the name content attribute."
-        self.html_element
-            .element
-            .get_attribute("name")
-            .unwrap_or_default()
-    }
-
-    /// <https://html.spec.whatwg.org/#dom-iframe-name>
-    pub(crate) fn set_name(&self, name: &str) {
-        // Step 1: "Set this's name content attribute to the given value."
-        self.html_element.element.set_attribute("name", name);
-    }
-
-    /// <https://html.spec.whatwg.org/#dom-dim-width>
-    pub(crate) fn width(&self) -> String {
-        // Step 1: "Return the value of the width content attribute."
-        self.html_element
-            .element
-            .get_attribute("width")
-            .unwrap_or_default()
-    }
-
-    /// <https://html.spec.whatwg.org/#dom-dim-width>
-    pub(crate) fn set_width(&self, width: &str) {
-        // Step 1: "Set this's width content attribute to the given value."
-        self.html_element.element.set_attribute("width", width);
-    }
-
-    /// <https://html.spec.whatwg.org/#dom-dim-height>
-    pub(crate) fn height(&self) -> String {
-        // Step 1: "Return the value of the height content attribute."
-        self.html_element
-            .element
-            .get_attribute("height")
-            .unwrap_or_default()
-    }
-
-    /// <https://html.spec.whatwg.org/#dom-dim-height>
-    pub(crate) fn set_height(&self, height: &str) {
-        // Step 1: "Set this's height content attribute to the given value."
-        self.html_element.element.set_attribute("height", height);
-    }
-}
