+++ webview/src/compositor.rs
@@ -1,6 +1,5 @@
 use anyrender::{PaintScene, Scene as RenderScene};
 use anyrender::recording::RenderCommand;
-use crate::log_iframe_debug;
 use ipc_messages::content::{FontTransportReceiver, FrameId, RecordedScene, ScrollOffset};
 use kurbo::{Affine, Shape};
 use peniko::{Color, Fill};
@@ -28,12 +27,6 @@ impl Compositor {
     }
 
     pub fn note_navigation_finalized(&mut self) {
-        log_iframe_debug(format!(
-            "note_navigation_finalized root_before={:?} committed_frames={} pending_frames={}",
-            self.root_frame_id.map(|id| id.0),
-            self.committed_frames.len(),
-            self.pending_frames.len()
-        ));
         self.pending_frames.clear();
         self.replace_root_on_next_paint = true;
     }
@@ -53,42 +46,21 @@ impl Compositor {
             scene,
         };
 
-        log_iframe_debug(format!(
-            "store_frame frame={} viewport={}x{} replace_root={} current_root={:?}",
-            frame_id.0,
-            viewport_width,
-            viewport_height,
-            self.replace_root_on_next_paint,
-            self.root_frame_id.map(|id| id.0)
-        ));
-
         if self.replace_root_on_next_paint {
             self.pending_frames.insert(frame_id, frame);
             if Self::frame_can_be_root(frame_id) {
                 self.root_frame_id = Some(frame_id);
                 self.committed_frames = std::mem::take(&mut self.pending_frames);
                 self.replace_root_on_next_paint = false;
-                log_iframe_debug(format!(
-                    "commit_new_root frame={} committed_frames={}",
-                    frame_id.0,
-                    self.committed_frames.len()
-                ));
             }
             return;
         }
 
         if self.root_frame_id.is_none() && Self::frame_can_be_root(frame_id) {
             self.root_frame_id = Some(frame_id);
-            log_iframe_debug(format!("select_initial_root frame={}", frame_id.0));
         }
 
         self.committed_frames.insert(frame_id, frame);
-        log_iframe_debug(format!(
-            "stored_committed_frame frame={} root={:?} committed_frames={}",
-            frame_id.0,
-            self.root_frame_id.map(|id| id.0),
-            self.committed_frames.len()
-        ));
     }
 
     pub fn committed_root_frame_id(&self) -> Option<FrameId> {
@@ -104,11 +76,6 @@ impl Compositor {
 
     pub fn compose_scene(&self, font_receiver: &FontTransportReceiver) -> Option<RenderScene> {
         let root_frame_id = self.root_frame_id?;
-        log_iframe_debug(format!(
-            "compose_scene root={} committed_frames={}",
-            root_frame_id.0,
-            self.committed_frames.len()
-        ));
         let mut stack = HashSet::from([root_frame_id]);
         self.compose_frame(root_frame_id, font_receiver, &mut stack)
     }
@@ -127,17 +94,7 @@ impl Compositor {
             match command {
                 RenderCommand::IframePlaceholder(placeholder) => {
                     let child_frame_id = FrameId(placeholder.frame_id);
-                    let clip_bounds = placeholder.clip.bounding_box();
                     if !stack.insert(child_frame_id) {
-                        log_iframe_debug(format!(
-                            "skip_placeholder_cycle parent={} child={} clip=({}, {}, {}x{})",
-                            frame_id.0,
-                            child_frame_id.0,
-                            clip_bounds.x0,
-                            clip_bounds.y0,
-                            clip_bounds.width(),
-                            clip_bounds.height()
-                        ));
                         continue;
                     }
 
@@ -148,15 +105,6 @@ impl Compositor {
                             .child_scene_transform(&placeholder.clip, child_frame_id)
                             .map(|transform| placeholder.transform * transform)
                             .unwrap_or(placeholder.transform);
-                        log_iframe_debug(format!(
-                            "compose_placeholder parent={} child={} clip=({}, {}, {}x{})",
-                            frame_id.0,
-                            child_frame_id.0,
-                            clip_bounds.x0,
-                            clip_bounds.y0,
-                            clip_bounds.width(),
-                            clip_bounds.height()
-                        ));
                         composed_scene.fill(
                             Fill::NonZero,
                             placeholder.transform,
@@ -167,16 +115,6 @@ impl Compositor {
                         composed_scene.push_clip_layer(placeholder.transform, &placeholder.clip);
                         composed_scene.append_scene(child_scene, child_transform);
                         composed_scene.pop_layer();
-                    } else {
-                        log_iframe_debug(format!(
-                            "missing_child_frame parent={} child={} clip=({}, {}, {}x{})",
-                            frame_id.0,
-                            child_frame_id.0,
-                            clip_bounds.x0,
-                            clip_bounds.y0,
-                            clip_bounds.width(),
-                            clip_bounds.height()
-                        ));
                     }
 
                     stack.remove(&child_frame_id);
@@ -191,30 +129,12 @@ impl Compositor {
     fn child_scene_transform(&self, clip: &impl Shape, child_frame_id: FrameId) -> Option<Affine> {
         let child_frame = self.committed_frames.get(&child_frame_id)?;
         if child_frame.viewport_width == 0 || child_frame.viewport_height == 0 {
-            log_iframe_debug(format!(
-                "child_scene_transform_zero_viewport frame={} viewport={}x{}",
-                child_frame_id.0,
-                child_frame.viewport_width,
-                child_frame.viewport_height
-            ));
             return None;
         }
 
         let clip_bounds = clip.bounding_box();
         let scale_x = clip_bounds.width() / f64::from(child_frame.viewport_width);
         let scale_y = clip_bounds.height() / f64::from(child_frame.viewport_height);
-        log_iframe_debug(format!(
-            "child_scene_transform frame={} child_viewport={}x{} clip=({}, {}, {}x{}) scale=({:.4}, {:.4})",
-            child_frame_id.0,
-            child_frame.viewport_width,
-            child_frame.viewport_height,
-            clip_bounds.x0,
-            clip_bounds.y0,
-            clip_bounds.width(),
-            clip_bounds.height(),
-            scale_x,
-            scale_y
-        ));
         Some(Affine::new([scale_x, 0.0, 0.0, scale_y, 0.0, 0.0]))
     }
 }
